+++ CHANGELOG.md
@@ -187,6 +187,8 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
   
 * Parenthesis are now inserted around all infix operations provided by diesel's `ExpressionMethods` traits
 
+* Queries containing a `distinct on` clause check now on compile time that a compatible order clause was set.
+
 ### Deprecated
 
 * `diesel_(prefix|postfix|infix)_operator!` have been deprecated. These macros